// packages/cantonjs-splice-validator/src/generated/walletExternal.types.ts
export interface components {
    schemas: {
// ...
        BaseGetTransferOfferStatusResponse: {
// ...
            status: string;
        };
// ...
        Contract: {
            contract_id?: components["schemas"]["ContractId"];
            template_id?: string;
            package_name?: string;
            payload?: {
                [key: string]: unknown;
            };
        } & {
            [key: string]: unknown;
        };
        ErrorResponse: {
            error?: string;
            message?: string;
            details?: {
                [key: string]: unknown;
            };
        } & {
            [key: string]: unknown;
        };
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
    listTransferOffers: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ListTransferOffersResponse"];
                };
            };
            404: components["responses"]["404"];
            500: components["responses"]["500"];
        };
    };
    createTransferOffer: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateTransferOfferRequest"];
            };
        };
        responses: {
            /** @description The transfer offer has been created. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["CreateTransferOfferResponse"];
                };
            };
            /** @description Invalid request, check the error response for details. */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description The submitter’s wallet could not be found. */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description A transfer offer with the same tracking id has been created. Check the status endpoint for the current status. */
            409: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description A transfer offer with the same tracking id is currently being processed, which may or may not succeed. Retry submitting the request with exponential back-off. */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            500: components["responses"]["500"];
        };
    };
    getTransferOfferStatus: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                tracking_id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description An offer with this tracking id is known. Check the response for its status. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetTransferOfferStatusResponse"];
                };
            };
            /**
             * @description No offer with this tracking id is known.
             *     Perhaps it has not yet been submitted or processed; or it has been submitted
             *     in the past before the current beginning of the wallet transaction log.
             */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
        };
    };
    createBuyTrafficRequest: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["CreateBuyTrafficRequest"];
            };
        };
        responses: {
            /** @description Request to buy traffic got created */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["CreateBuyTrafficRequestResponse"];
                };
            };
            /** @description Request was invalid, adjust parameters */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description A request to buy traffic with the same tracking id has been created. Check the status endpoint for the current status. */
            409: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description A request to buy traffic with the same tracking id is currently being processed. Check the status endpoint and resubmit if it returns 404. */
            429: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
            /** @description Internal server error that requires operator investigation. Retrying for a small number of times with exponential back-off MAY work. */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
        };
    };
    getBuyTrafficRequestStatus: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                tracking_id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description A request to buy traffic with this tracking id has been submitted before, check the response for details. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetBuyTrafficRequestStatusResponse"];
                };
            };
            /** @description No request with this tracking id was found. */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ErrorResponse"];
                };
            };
        };
    };
}